+++ noodles-bam/src/record.rs
@@ -53,7 +53,7 @@ pub struct Record {
     pub(crate) next_pos: i32,
     pub(crate) tlen: i32,
     pub(crate) read_name: Vec<u8>,
-    pub(crate) cigar: Vec<u8>,
+    pub(crate) cigar: Vec<u32>,
     pub(crate) seq: Vec<u8>,
     pub(crate) qual: Vec<u8>,
     pub(crate) data: Vec<u8>,
@@ -73,7 +73,7 @@ impl Record {
             + mem::size_of::<i32>() // next_pos
             + mem::size_of::<i32>() // tlen
             + self.read_name.len()
-            + self.cigar.len()
+            + (mem::size_of::<u32>() * self.cigar.len())
             + self.seq.len()
             + self.qual.len()
             + self.data.len()
@@ -351,9 +351,9 @@ mod tests {
             next_pos: 61152,
             tlen: 166,
             read_name: b"r0\x00".to_vec(),
-            cigar: vec![0x40, 0x00, 0x00, 0x00], // 4M
-            seq: vec![0x18, 0x42],               // ATGC
-            qual: vec![0x1f, 0x1d, 0x1e, 0x20],  // @>?A
+            cigar: vec![0x00000040],            // 4M
+            seq: vec![0x18, 0x42],              // ATGC
+            qual: vec![0x1f, 0x1d, 0x1e, 0x20], // @>?A
             data: vec![
                 0x4e, 0x4d, 0x43, 0x00, // NM:i:0
                 0x50, 0x47, 0x5a, 0x53, 0x4e, 0x41, 0x50, 0x00, // PG:Z:SNAP
@@ -438,7 +438,7 @@ mod tests {
     #[test]
     fn test_cigar() {
         let record = build_record();
-        assert_eq!(*record.cigar(), [0x40, 0x00, 0x00, 0x00]);
+        assert_eq!(*record.cigar(), [0x00000040]);
     }
 
     #[test]