+++ noodles-bam/src/record/cigar/ops.rs
@@ -1,18 +1,15 @@
-use std::{io, mem};
+use std::{io, slice};
 
 use super::Op;
 
 /// An iterator over the operations of a CIGAR.
 ///
 /// This is created by calling [`super::Cigar::ops`].
-pub struct Ops<'a> {
-    cigar: &'a [u8],
-    i: usize,
-}
+pub struct Ops<'a>(slice::Iter<'a, u32>);
 
 impl<'a> Ops<'a> {
-    pub(crate) fn new(cigar: &'a [u8]) -> Self {
-        Self { cigar, i: 0 }
+    pub(crate) fn new(cigar: &'a [u32]) -> Self {
+        Self(cigar.iter())
     }
 }
 
@@ -20,22 +17,9 @@ impl<'a> Iterator for Ops<'a> {
     type Item = io::Result<Op>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let size = mem::size_of::<u32>();
-        let start = self.i * size;
-
-        if start < self.cigar.len() {
-            let data = &self.cigar[start..];
-
-            let option = Op::try_from(data)
-                .map(Some)
-                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-                .transpose();
-
-            self.i += 1;
-
-            option
-        } else {
-            None
-        }
+        self.0
+            .next()
+            .copied()
+            .map(|n| Op::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)))
     }
 }